Refactored film-details component to separate files
