+++ src/components/film-details/genre-markup.js
@@ -1,0 +1,9 @@
+/**
+ * Generates genres markup
+ * @param {Set<*>} genres - Set of genres
+ * @return {String} - genres markup
+ */
+const generateGenreMarkup = (genres) => [...genres]
+  .map((genre) => `<span class="film-details__genre">${genre}</span>`);
+
+export {generateGenreMarkup};
