+++ src/components/film-details/index.js
@@ -1,0 +1,16 @@
+import Component from "../component";
+import {template} from "./template";
+
+export default class FilmDetails extends Component {
+  constructor(film) {
+    super(template(film));
+  }
+
+  /**
+   * Returns components close element
+   * @return {HTMLElement} - close element
+   */
+  getCloseElement() {
+    return this.getElement().querySelector(`.film-details__close-btn`);
+  }
+}
